remove peer dependencies from runtime dependencies
The `peerDependencies` and `dependencies` are conflicting with one another,
resulting in installations of this component breaking when `tsc` is run
against the project. TypeScript is mistakenly using the wrong type definitions
for the dependencies here after this package was introduced, so these can't
live here in the runtime dependencies, they have to be specified in peer deps
and copied over into dev deps so you can run tests against them.

I had some problems with linting when I tried to commit locally, so this was
committed with `--no-verify`, which I can fix up whenever I figure out why
`@tsconfig/docusaurus` isn't installed and why installing it causes everything
to break, since that seems to be what ESLint is balking over.

Resolves: #18